uefi: Make `exit_boot_services` handle memory map allocation
Exiting boot services is a one-way operation, which is why `exit_boot_services`
takes `self` rather than a reference. If it succeeds then you get back a
`SystemTable<Runtime>`, but even if it fails you can't continue to use boot
services except for a very limited subset. Specifically, in UEFI 2.8 and
earlier, only `get_memory_map` and `exit_boot_services` are allowed. Starting in
UEFI 2.9 other memory allocation functions may also be called.

The previous API had the caller pass in a buffer to hold the memory map,
requiring the caller to ensure that the buffer was big enough. On failure, the
caller would have no way to

This resolves an issue with the previous API: the function could fail due to the
memory map being too small, but the caller no longer had access to
`SystemTable<Boot>` and so they couldn't allocate a bigger buffer.

The new API takes no arguments and handles the allocation internally,
automatically increasing the size of the memory map buffer on each retry. The
lifetime of the returned memory map is now `'static`, which makes sense since
the memory can no longer be free'd through boot services.