Move Status, Error, Result, and ResultExt to uefi-raw
Only the `Status` type really belongs in uefi-raw, but we have a lot of
conversion methods between the various types that would be less convenient if
they weren't in the same crate. Consider, for example, the impl of
`From<Status>` for `Result`. It would not be possible to impl that in `uefi`
with `Status` was in `uefi-raw`.